requirements.txt
argparse
dnspython
requests
pyfiglet
termcolor
colorama
virtualenv